docs: #276 migrate data on firebase
