+++ mw-webapp/src/logic/aboutProjectPage/AboutProjectPageContent.json
@@ -99,7 +99,7 @@
         "en": "I like your project, how can I help you?"
       },
       "description": {
-        "ru": "Мы рады рассмотреть любое предложение о помощи, вышлите описание на почту victorveretennikov58@gmail.com (с темой - 'Mastersway:cooperation'). Так же вы можете отправить небольшое пожертвование на крипто кошельки: \nBitcoin: bc1qwemlhwklwsfl3xejkdg7w6s02x7s99eyy8vgwj \netherium: 0x24e32141E945D88a703051605f568C489ea89081 \nteher: 0x24e32141E945D88a703051605f568C489ea89081",
+        "ru": "Мы рады рассмотреть любое предложение о помощи, вышлите описание на почту victorveretennikov58@gmail.com (с темой - 'Mastersway:cooperation'). Так же вы можете отправить небольшое пожертвование на крипто кошельки: \nBitcoin: bc1qwemlhwklwsfl3xejkdg7w6s02x7s99eyy8vgwj \netherium: 0x24e32141E945D88a703051605f568C489ea89081 \ntether: 0x24e32141E945D88a703051605f568C489ea89081",
         "en": "We are happy to consider any offer of help, send a description to victorveretennikov58@gmail.com (with the subject - 'Mastersway:cooperation'). You can also send a small donation to crypto wallets: \nBitcoin: bc1qwemlhwklwsfl3xejkdg7w6s02x7s99eyy8vgwj \netherium: 0x24e32141E945D88a703051605f568C489ea89081 \nteher: 0x24e32141E945D88a703051605f568C489ea89081"
       }
     }
